+++ pkg/schema/int32.go
@@ -64,7 +64,7 @@ func (s *Int32Schema) Validate(model *ModelSchema) error {
 
 type Int32ArraySchema struct {
 	Name           string                     `hcl:"name,label"`
-	InitialSize    uint64                     `hcl:"initial_size,attr"`
+	InitialSize    uint32                     `hcl:"initial_size,attr"`
 	Accessor       *bool                      `hcl:"accessor,optional"`
 	LimitValidator *Int32LimitValidatorSchema `hcl:"limit_validator,block"`
 }
